Make the settings page look different
