+++ SIT.Manager/Views/MainView.axaml
@@ -91,9 +91,7 @@
 
 			<ui:NavigationView.FooterMenuItems>
 				<ui:NavigationViewItem IconSource="Help" Content="Help" Tag="Help"/>
-				<ui:NavigationViewItem IconSource="Settings" Content="{DynamicResource LinuxSettingsPageTitle}" Tag="SIT.Manager.Views.LinuxSettingsPage" IsVisible="{OnPlatform Default=False, Linux=True}"/>
-				<ui:NavigationViewItem IconSource="Settings" Content="{DynamicResource SettingsTitle}" Tag="SIT.Manager.Views.SettingsPage"/>
-			</ui:NavigationView.FooterMenuItems>		
+				<ui:NavigationViewItem IconSource="Settings" Content="{DynamicResource SettingsTitle}" Tag="SIT.Manager.Views.SettingsPage"/>			</ui:NavigationView.FooterMenuItems>		
 
 			<Grid RowDefinitions="*,Auto">
 				<ui:Frame Name="ContentFrame"
