+++ SIT.Manager/Views/Settings/LauncherView.axaml.cs
@@ -1,0 +1,14 @@
+using Microsoft.Extensions.DependencyInjection;
+using SIT.Manager.Controls;
+using SIT.Manager.ViewModels.Settings;
+
+namespace SIT.Manager.Views.Settings;
+
+public partial class LauncherView : ActivatableUserControl
+{
+    public LauncherView()
+    {
+        InitializeComponent();
+        DataContext = App.Current.Services.GetService<LauncherViewModel>();
+    }
+}
