+++ SIT.Manager/Views/Settings/LinuxView.axaml.cs
@@ -1,0 +1,13 @@
+using Microsoft.Extensions.DependencyInjection;
+using SIT.Manager.Controls;
+using SIT.Manager.ViewModels.Settings;
+
+namespace SIT.Manager.Views.Settings;
+public partial class LinuxView : ActivatableUserControl
+{
+    public LinuxView()
+    {
+        InitializeComponent();
+        DataContext = App.Current.Services.GetService<LinuxViewModel>();
+    }
+}
