+++ SIT.Manager/Views/Settings/SptAkiView.axaml
@@ -1,0 +1,79 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+			 xmlns:ui="using:FluentAvalonia.UI.Controls"
+			 xmlns:vm="clr-namespace:SIT.Manager.ViewModels.Settings"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="SIT.Manager.Views.Settings.SptAkiView"
+			 x:DataType="vm:SptAkiViewModel">
+	<StackPanel>
+		<Border Classes="StandardFrame">
+			<StackPanel>
+				<StackPanel Orientation="Horizontal">
+					<TextBlock Text="{DynamicResource SettingsSPTAKIPathTitle}"  VerticalAlignment="Center"/>
+					<TextBox Watermark="{DynamicResource SettingsSPTAKIInstallPathPlaceholder}"
+							 Margin="5"
+							 MinWidth="397"
+							 MaxWidth="425"
+							 IsReadOnly="True"
+							 Text="{Binding Config.AkiServerPath}"/>
+					<Button Name="ChangeAkiServerPath"
+							Content="{DynamicResource SettingsEFTSettingsButtonChangeTitle}"
+							Margin="5"
+							ToolTip.Tip="{DynamicResource SettingsSPTAKIInstallPathToolTip}"
+							Command="{Binding ChangeAkiServerLocationCommand}"/>
+				</StackPanel>
+
+				<TextBlock Margin="0,10,0,0">
+					<Run Text="{DynamicResource SettingsSPTAKIVersionTitle}"/>
+					<Run Text=" "/>
+					<Run Text="{Binding Config.SptAkiVersion}"/>
+				</TextBlock>
+
+				<TextBlock Margin="0,10,0,0">
+					<Run Text="{DynamicResource SettingsSITModVersionTitle}"/>
+					<Run Text=" "/>
+					<Run Text="{Binding Config.SitModVersion}"/>
+				</TextBlock>
+			</StackPanel>
+		</Border>
+		
+		<!--Server Console-->
+		<TextBlock Text="{DynamicResource SettingsServerConsoleTitle}" Classes="FrameHeading"/>
+		<Border Classes="StandardFrame">
+			<StackPanel>
+				<StackPanel Orientation="Horizontal">
+					<TextBlock Text="{DynamicResource SettingsServerConsoleFontFamilyTitle}" Margin="0,0,10,0" VerticalAlignment="Center"/>
+					<ComboBox x:Name="fontComboBox"
+							  Margin="0,0,10,0"
+							  VerticalAlignment="Center"
+							  VerticalContentAlignment="Center"
+							  SelectedIndex="0"
+							  MaxDropDownHeight="300"
+							  ItemsSource="{Binding InstalledFonts}"
+							  SelectedItem="{Binding SelectedConsoleFontFamily}">
+						<ComboBox.ItemTemplate>
+							<DataTemplate>
+								<TextBlock Text="{Binding}"
+										   FontFamily="{Binding}"
+										   VerticalAlignment="Center"/>
+							</DataTemplate>
+						</ComboBox.ItemTemplate>
+					</ComboBox>
+				</StackPanel>
+				<StackPanel Orientation="Horizontal">
+					<TextBlock Text="{DynamicResource SettingsServerConsoleFontColorTitle}" Margin="0,10,10,0" VerticalAlignment="Center"/>
+					<ui:ColorPickerButton Name="PickerButton"
+										  Margin="0,10,0,0"
+										  IsCompact="True"
+										  IsMoreButtonVisible="False"
+										  UseColorPalette="False"
+										  UseColorWheel="False"
+										  UseSpectrum="True"
+										  Color="{Binding Config.ConsoleFontColor, Mode=TwoWay}" />
+				</StackPanel>
+			</StackPanel>
+		</Border>
+	</StackPanel>
+</UserControl>
