+++ SIT.Manager/Views/Settings/SptAkiView.axaml.cs
@@ -1,0 +1,14 @@
+using Microsoft.Extensions.DependencyInjection;
+using SIT.Manager.Controls;
+using SIT.Manager.ViewModels.Settings;
+
+namespace SIT.Manager.Views.Settings;
+
+public partial class SptAkiView : ActivatableUserControl
+{
+    public SptAkiView()
+    {
+        InitializeComponent();
+        DataContext = App.Current.Services.GetService<SptAkiViewModel>();
+    }
+}
